% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_files.R
\name{choose_files}
\alias{choose_files}
\title{Choose files}
\usage{
choose_files(
  input_folder,
  l0_file_path,
  reprocess = F,
  start_date = NULL,
  end_date = NULL,
  files_to_exclude = NULL
)
}
\arguments{
\item{input_folder}{Folder where input .dat files are stored}

\item{l0_file_path}{L0 file path}

\item{reprocess}{Whether or not to re-process calcualted fluxes (T/F)}

\item{start_date}{Start date of files to process}

\item{end_date}{End date of files to process}

\item{files_to_exclude}{Vector of file names to exclude}
}
\value{
List of files
}
\description{
Choose files
}
