% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_slopes.R
\name{combine_slopes}
\alias{combine_slopes}
\title{Combine slopes}
\usage{
combine_slopes(new, old)
}
\arguments{
\item{new}{new (re-processed) slopes}

\item{old}{older slope file}
}
\value{
combined slopes
}
\description{
Combine slopes
}
