% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_small_chapada}
\alias{data_small_chapada}
\title{Formatted data from Chapada}
\format{
\subsection{\code{data_small_chapada}}{

A data frame with 80,130 rows and 7 columns:
\describe{
\item{location}{Site name}
\item{TIMESTAMP}{Timestamp}
\item{CH4d_ppm}{Methane concentration (ppm)}
\item{CO2d_ppm}{Carbon dioxide concentration (ppm)}
\item{Chamber}{Sampling chamber}
\item{Manifold_Timer}{Time since switching to this valve}
\item{Flux_Status}{Status of chamber (e.g., open/closed)}
...
}
}
}
\source{
\url{https://smithsonian.github.io/Chapada_Stem/}
}
\usage{
data_small_chapada
}
\description{
Formatted data from Chapada
}
\keyword{datasets}
