% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_small_genx}
\alias{data_small_genx}
\title{Data from GENX and Chapada flux chambers}
\format{
\subsection{\code{data_small_genx}}{

A data frame with 73,800 rows and 7 columns:
\describe{
\item{TIMESTAMP}{Timestamp}
\item{CH4d_ppm}{Methane concentration (ppm)}
\item{CO2d_ppm}{Carbon dioxide concentration (ppm)}
\item{N2Od_ppm}{Nitrous oxide concentration (ppm)}
\item{MIU_VALVE}{Sampling valve (chamber)}
\item{Manifold_Timer}{Time since switching to this valve}
\item{Format}{Data format (only NEW works for this processing code)}
...
}
}
}
\source{
\url{https://serc.si.edu/labs/biogeochemistry-projects/genx}
}
\usage{
data_small_genx
}
\description{
Formatted data from GENX
}
\keyword{datasets}
