% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_co2.R
\name{estimate_params_moving_window}
\alias{estimate_params_moving_window}
\title{Moving-window parameter estimation per chamber}
\usage{
estimate_params_moving_window(
  dt_ch,
  window_days = 30,
  step_days = 1,
  par_night_thresh = 5,
  Tref = 10
)
}
\arguments{
\item{dt_ch}{}

\item{window_days}{}

\item{step_days}{}

\item{par_night_thresh}{}

\item{Tref}{}
}
\value{
Dataframe with estimated parameters
}
\description{
Moving-window parameter estimation per chamber
}
