% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fluxes_by_SE.R
\name{filter_fluxes_by_SE}
\alias{filter_fluxes_by_SE}
\title{Remove fluxes with low SE}
\usage{
filter_fluxes_by_SE(seasonal_fluxes, group_cols = NULL)
}
\arguments{
\item{seasonal_fluxes}{Seasonal flux file}

\item{group_cols}{Columns to group by. Must be in data and maintenance log}
}
\value{
seasonal_fluxes dataframe with removed values and updated flags
}
\description{
Remove fluxes with low SE
}
