% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_co2.R
\name{fit_q10_lm}
\alias{fit_q10_lm}
\title{Fit Q10}
\usage{
fit_q10_lm(dt_night, Tref = 10, min_night = 5)
}
\arguments{
\item{dt_night}{}

\item{Tref}{}

\item{min_night}{}
}
\value{
List of fitted parameters
}
\description{
Fit Q10
}
