% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_recent_raw.R
\name{generate_recent_raw}
\alias{generate_recent_raw}
\title{Generate recent raw data}
\usage{
generate_recent_raw(
  data_small,
  today = Sys.Date(),
  group_cols = NULL,
  days = 7
)
}
\arguments{
\item{data_small}{dataset. Must contain columns for MIU_VALVE, TIMESTAMP,
and CH4d_ppm/CO2d_ppm/N2Od_ppm}

\item{today}{Latest date of flux data requested}

\item{group_cols}{Optional grouping variable}

\item{days}{Number of days of data}
}
\value{
Recent raw dataset
}
\description{
Generate an output file of recent raw data
}
