% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_fun.R
\name{group_fun}
\alias{group_fun}
\title{Assign groups for separate readings}
\usage{
group_fun(Chamber)
}
\arguments{
\item{Chamber}{Chamber number}
}
\value{
group number
}
\description{
Internal utility for assigning fluxes to groups based on chamber number
}
