% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_loggernet_flux_data.R
\name{load_loggernet_flux_data}
\alias{load_loggernet_flux_data}
\title{Load and combine data}
\usage{
load_loggernet_flux_data(files, format)
}
\arguments{
\item{files}{list of file paths}

\item{format}{file format ("GENX" or "Chapada")}
}
\value{
Combined and formatted data
}
\description{
This file takes in a list of file paths and loads/formats the data for
analysis. The formatting done here is specific to the loggernet file
formats for the GENX and Chapada projects.
}
