% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_single_file.R
\name{load_single_file_chapada}
\alias{load_single_file_chapada}
\title{Load data- ChapadaSTEM}
\usage{
load_single_file_chapada(file)
}
\arguments{
\item{file}{file path}
}
\value{
formatted file
}
\description{
Load data- ChapadaSTEM
}
