% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_single_file.R
\name{load_single_file_genx}
\alias{load_single_file_genx}
\title{Load data- GENX}
\usage{
load_single_file_genx(file)
}
\arguments{
\item{file}{File path}
}
\value{
Formatted file
}
\description{
Load data- GENX
}
