% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_maintenance_log.R
\name{add_maintenance_log}
\alias{add_maintenance_log}
\title{Add maintenance log}
\usage{
add_maintenance_log(slopes, gs_url, group_cols = NULL)
}
\arguments{
\item{slopes}{Calculated slopes}

\item{gs_url}{Google sheets URL for maintenance log}

\item{group_cols}{Columns to group by. Must be in data and maintenance log}
}
\value{
Updated slopes with maintenance applied
}
\description{
This function uses a maintenance log on Google Sheets to identify periods
of time in a flux dataframe that should be flagged and potentially removed
(set to NA).
}
