% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flux.R
\name{calculate_flux}
\alias{calculate_flux}
\title{calculate_flux}
\usage{
calculate_flux(data_small, cutoff_start, cutoff_end, group_cols = NULL)
}
\arguments{
\item{data_small}{dataset. Must contain columns for MIU_VALVE, TIMESTAMP,
and CH4d_ppm/CO2d_ppm/N2Od_ppm}

\item{cutoff_start}{Amount of time to remove after the start of the flux
interval (in seconds)}

\item{cutoff_end}{End of flux, as determined from the start of the flux
interval (in seconds)}

\item{group_cols}{Optional grouping variable. See details}
}
\description{
Calculate linear rate of change in CO2, CH4, and/or N2O for each flux interval
}
\details{
This function calculates the linear rate of change in gas (CO2, CH4,
and/or N2O) concentrations. The function requires a dataframe of input data
that contains columns for gas concentrations (in ppm), timestamps, and
a column with the chamber index. By default, fluxes are grouped by
timestamp and chamber index, but additional grouping columns can be
identified in group_cols (e.g., if data from multiple gas analyzers are
combined in data_small)
}
