% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_fun.R
\name{group_fun}
\alias{group_fun}
\title{Assign groups for separate readings}
\usage{
group_fun(MIU_VALVE)
}
\arguments{
\item{MIU_VALVE}{Chamber number}
}
\value{
group number
}
\description{
Assign groups for separate readings
}
